build: enable libseat support by default
Now that launcher-direct and weston-launch are gone, libseat takes their
place.

Enable libseat support by default to give users a hint in case they miss
either of those.

People who used to get launcher-logind when libseat support was disabled
will now be using logind through libseat. This should not cause any
regressions, and if it does, we want to hear about them, because the
separate logind-launcher is also planned to be deprecated in the future.

Disabling logind-launcher by default is left for when it actually gets
deprecated.

In case someone does not have libseat available but do have logind
running, they can just disable libseat support.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>